Oslo-messaging-zmq-receiver cannot recive any messages
In Icehouse b3 and rc1, when we use impl_zmq as
rpc_backend in nova, nova cannot send any messages
throught impl_zmq to oslo-messaging-zmq-receiver.
The reason is we send two params to serialize_msg
method but it only can recive one.To fix this
problem, I removed the second param which not
used when we call the method serialize_msg.

Change-Id: I2e7546147affd9493f425fd879169932ef206d48
Closes-bug: #1301132